+++ src/Engine.ts
@@ -1,6 +1,8 @@
 import { QlikRepositoryClient } from "./main";
+import { URLBuild } from "./util/generic";
 import { modifiedDateTime } from "./util/generic";
 
+import { ISelection } from "./types/interfaces";
 import { ITagCondensed } from "./Tag";
 import { ICustomPropertyCondensed } from "./CustomProperty";
 import { IServerNodeConfigurationCondensed } from "./Node";
@@ -94,6 +96,7 @@ export interface IClassEngine {
   getAll(): Promise<IEngineCondensed[]>;
   getFilter(filter: string): Promise<IEngine[]>;
   getValid(arg?: IEngineGetValid): Promise<IEngineGetValidResult[]>;
+  select(filter?: string): Promise<ISelection>;
   update(arg: IEngineUpdate): Promise<IEngine>;
 }
 
@@ -104,7 +107,7 @@ export class Engine implements IClassEngine {
   }
 
   public async get(id: string) {
-    if (!id) throw new Error(`engineGet: "id" parameter is required`);
+    if (!id) throw new Error(`engine.get: "id" parameter is required`);
     return await this.repoClient
       .Get(`engineservice`)
       .then((res) => res.data as IEngine);
@@ -137,7 +140,7 @@ export class Engine implements IClassEngine {
   }
 
   public async getFilter(filter: string) {
-    if (!filter) throw new Error(`engineGetFilter: "filter" is required`);
+    if (!filter) throw new Error(`engine.getFilter: "filter" is required`);
 
     let baseUrl = `engineservice/full`;
 
@@ -146,27 +149,42 @@ export class Engine implements IClassEngine {
       .then((res) => res.data as IEngine[]);
   }
 
+  public async select(filter?: string) {
+    const urlBuild = new URLBuild(`selection/engineservice`);
+    urlBuild.addParam("filter", filter);
+
+    return await this.repoClient
+      .Post(urlBuild.getUrl(), {})
+      .then((res) => res.data as ISelection);
+  }
+
   // REVIEW: double check the whole logic here
   public async update(arg: IEngineUpdate) {
-    if (!arg.id) throw new Error(`engineUpdate: "id" is required`);
+    if (!arg.id) throw new Error(`engine.update: "id" is required`);
 
     let engine = await this.get(arg.id);
 
     if (arg.workingSetSizeLoPct) {
       if (arg.workingSetSizeLoPct < 0 || arg.workingSetSizeLoPct > 100)
-        throw new Error("workingSetSizeLoPct must be between 0 and 100");
+        throw new Error(
+          "engine.update: workingSetSizeLoPct must be between 0 and 100"
+        );
       engine.settings.workingSetSizeLoPct = arg.workingSetSizeLoPct;
     }
 
     if (arg.workingSetSizeHiPct) {
       if (arg.workingSetSizeHiPct < 0 || arg.workingSetSizeHiPct > 100)
-        throw new Error("workingSetSizeHiPct must be between 0 and 100");
+        throw new Error(
+          "engine.update: workingSetSizeHiPct must be between 0 and 100"
+        );
       engine.settings.workingSetSizeHiPct = arg.workingSetSizeHiPct;
     }
 
     if (arg.cpuThrottlePercentage) {
       if (arg.cpuThrottlePercentage < 0 || arg.cpuThrottlePercentage > 100)
-        throw new Error("cpuThrottlePercentage must be between 0 and 100");
+        throw new Error(
+          "engine.update: cpuThrottlePercentage must be between 0 and 100"
+        );
       engine.settings.cpuThrottlePercentage = arg.cpuThrottlePercentage;
     }
 
@@ -177,7 +195,7 @@ export class Engine implements IClassEngine {
         arg.workingSetSizeMode != "HardMaxLimit"
       )
         throw new Error(
-          `Engine config working set valid values are: IgnoreMaxLimit, SoftMaxLimit, HardMaxLimit`
+          `engine.update: Engine config working set valid values are: IgnoreMaxLimit, SoftMaxLimit, HardMaxLimit`
         );
 
       if (arg.workingSetSizeMode == "IgnoreMaxLimit")
@@ -192,7 +210,9 @@ export class Engine implements IClassEngine {
 
     if (arg.coresToAllocate) {
       if (arg.coresToAllocate < 0 || arg.coresToAllocate > 256)
-        throw new Error("coresToAllocate must be between 0 and 256");
+        throw new Error(
+          "engine.update: coresToAllocate must be between 0 and 256"
+        );
 
       let mask = [0, 0, 0, 0, 0, 0, 0, 0];
       let bin = "".padEnd(arg.coresToAllocate, "1").padStart(256, "0");
@@ -240,7 +260,9 @@ export class Engine implements IClassEngine {
         arg.auditActivityLogVerbosity < 0 ||
         arg.auditActivityLogVerbosity > 5
       )
-        throw new Error("auditActivityLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: auditActivityLogVerbosity must be between 0 and 5"
+        );
       engine.settings.auditActivityLogVerbosity = arg.auditActivityLogVerbosity;
     }
 
@@ -249,13 +271,17 @@ export class Engine implements IClassEngine {
         arg.auditSecurityLogVerbosity < 0 ||
         arg.auditSecurityLogVerbosity > 5
       )
-        throw new Error("auditSecurityLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: auditSecurityLogVerbosity must be between 0 and 5"
+        );
       engine.settings.auditSecurityLogVerbosity = arg.auditSecurityLogVerbosity;
     }
 
     if (arg.systemLogVerbosity) {
       if (arg.systemLogVerbosity < 0 || arg.systemLogVerbosity > 5)
-        throw new Error("systemLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: systemLogVerbosity must be between 0 and 5"
+        );
       engine.settings.systemLogVerbosity = arg.systemLogVerbosity;
     }
 
@@ -264,7 +290,9 @@ export class Engine implements IClassEngine {
         arg.externalServicesLogVerbosity < 0 ||
         arg.externalServicesLogVerbosity > 5
       )
-        throw new Error("externalServicesLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: externalServicesLogVerbosity must be between 0 and 5"
+        );
       engine.settings.externalServicesLogVerbosity =
         arg.externalServicesLogVerbosity;
     }
@@ -274,50 +302,66 @@ export class Engine implements IClassEngine {
         arg.qixPerformanceLogVerbosity < 0 ||
         arg.qixPerformanceLogVerbosity > 5
       )
-        throw new Error("qixPerformanceLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: qixPerformanceLogVerbosity must be between 0 and 5"
+        );
       engine.settings.qixPerformanceLogVerbosity =
         arg.qixPerformanceLogVerbosity;
     }
 
     if (arg.serviceLogVerbosity) {
       if (arg.serviceLogVerbosity < 0 || arg.serviceLogVerbosity > 5)
-        throw new Error("serviceLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: serviceLogVerbosity must be between 0 and 5"
+        );
       engine.settings.serviceLogVerbosity = arg.serviceLogVerbosity;
     }
 
     if (arg.httpTrafficLogVerbosity) {
       if (arg.httpTrafficLogVerbosity < 0 || arg.httpTrafficLogVerbosity > 5)
-        throw new Error("httpTrafficLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: httpTrafficLogVerbosity must be between 0 and 5"
+        );
       engine.settings.httpTrafficLogVerbosity = arg.httpTrafficLogVerbosity;
     }
 
     if (arg.auditLogVerbosity) {
       if (arg.auditLogVerbosity < 0 || arg.auditLogVerbosity > 5)
-        throw new Error("auditLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: auditLogVerbosity must be between 0 and 5"
+        );
       engine.settings.auditLogVerbosity = arg.auditLogVerbosity;
     }
 
     if (arg.trafficLogVerbosity) {
       if (arg.trafficLogVerbosity < 0 || arg.trafficLogVerbosity > 5)
-        throw new Error("trafficLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: trafficLogVerbosity must be between 0 and 5"
+        );
       engine.settings.trafficLogVerbosity = arg.trafficLogVerbosity;
     }
 
     if (arg.sessionLogVerbosity) {
       if (arg.sessionLogVerbosity < 0 || arg.sessionLogVerbosity > 5)
-        throw new Error("sessionLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: sessionLogVerbosity must be between 0 and 5"
+        );
       engine.settings.sessionLogVerbosity = arg.sessionLogVerbosity;
     }
 
     if (arg.performanceLogVerbosity) {
       if (arg.performanceLogVerbosity < 0 || arg.performanceLogVerbosity > 5)
-        throw new Error("performanceLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: performanceLogVerbosity must be between 0 and 5"
+        );
       engine.settings.performanceLogVerbosity = arg.performanceLogVerbosity;
     }
 
     if (arg.sseLogVerbosity) {
       if (arg.sseLogVerbosity < 0 || arg.sseLogVerbosity > 5)
-        throw new Error("auditActivityLogVerbosity must be between 0 and 5");
+        throw new Error(
+          "engine.update: auditActivityLogVerbosity must be between 0 and 5"
+        );
       engine.settings.sseLogVerbosity = arg.sseLogVerbosity;
     }