+++ src/Extension.ts
@@ -2,14 +2,14 @@ import { QlikRepositoryClient } from "./main";
 import { UpdateCommonProperties } from "./util/UpdateCommonProps";
 import { URLBuild } from "./util/generic";
 
+import { ISelection } from "./types/interfaces";
 import { ICustomPropertyCondensed } from "./CustomProperty";
 import { ITagCondensed } from "./Tag";
 
 import {
-  IHttpStatus,
+  IOwner,
   IEntityRemove,
   IFileExtensionWhiteListCondensed,
-  IOwner,
   IStaticContentReferenceCondensed,
 } from "./types/interfaces";
 
@@ -48,6 +48,8 @@ export interface IClassExtension {
   getFilter(filter: string, full?: boolean): Promise<IExtensionCondensed[]>;
   import(arg: IExtensionImport): Promise<IExtension>;
   remove(id: string): Promise<IEntityRemove>;
+  removeFilter(filter: string): Promise<IEntityRemove[]>;
+  select(filter?: string): Promise<ISelection>;
   update(arg: IExtensionUpdate): Promise<IExtension>;
 }
 
@@ -58,7 +60,7 @@ export class Extension implements IClassExtension {
   }
 
   public async get(id: string) {
-    if (!id) throw new Error(`extensionGet: "id" parameter is required`);
+    if (!id) throw new Error(`extension.get: "id" parameter is required`);
     return await this.repoClient
       .Get(`extension/${id}`)
       .then((res) => res.data as IExtension);
@@ -72,7 +74,7 @@ export class Extension implements IClassExtension {
 
   public async getFilter(filter: string) {
     if (!filter)
-      throw new Error(`extensionGetFilter: "filter" parameter is required`);
+      throw new Error(`extension.getFilter: "filter" parameter is required`);
 
     return await this.repoClient
       .Get(`extension?filter=(${encodeURIComponent(filter)})`)
@@ -80,15 +82,42 @@ export class Extension implements IClassExtension {
   }
 
   public async remove(id: string) {
-    if (!id) throw new Error(`extensionRemove: "id" parameter is required`);
+    if (!id) throw new Error(`extension.remove: "id" parameter is required`);
 
     return await this.repoClient.Delete(`extension/${id}`).then((res) => {
       return { id, status: res.status } as IEntityRemove;
     });
   }
 
+  public async removeFilter(filter: string) {
+    if (!filter)
+      throw new Error(`extension.removeFilter: "filter" parameter is required`);
+
+    const tags = await this.getFilter(filter).then((t: IExtension[]) => {
+      if (t.length == 0)
+        throw new Error(`extension.removeFilter: filter query return 0 items`);
+
+      return t;
+    });
+    return await Promise.all<IEntityRemove>(
+      tags.map((tag: IExtension) => {
+        return this.remove(tag.id);
+      })
+    );
+  }
+
+  public async select(filter?: string) {
+    const urlBuild = new URLBuild(`selection/extension`);
+    urlBuild.addParam("filter", filter);
+
+    return await this.repoClient
+      .Post(urlBuild.getUrl(), {})
+      .then((res) => res.data as ISelection);
+  }
+
   public async update(arg: IExtensionUpdate) {
-    if (!arg.id) throw new Error(`extensionUpdate: "id" parameter is required`);
+    if (!arg.id)
+      throw new Error(`extension.update: "id" parameter is required`);
 
     let extension = await this.get(arg.id);
 
@@ -102,7 +131,7 @@ export class Extension implements IClassExtension {
 
   public async import(arg: IExtensionImport) {
     if (!arg.file)
-      throw new Error(`extensionImport: "file" parameter is required`);
+      throw new Error(`extension.import: "file" parameter is required`);
 
     const urlBuild = new URLBuild(`extension/upload`);
     if (arg.password) urlBuild.addParam("password", arg.password);